# Conflicts:
#	src/main/java/org/math3d/Vector4.java
#	target/classes/module-info.class
#	target/classes/org/math/Line.class
#	target/classes/org/math/Vector.class
#	target/classes/org/math3d/Vector3.class
#	target/classes/org/math3d/Vector4.class
#	target/classes/org/utils/Utils.class
#	target/test-classes/org/math/VectorTest.class

+++ src/main/java/org/utils/Utils.java
@@ -26,7 +26,7 @@ public class Utils {
      * Returns true if a double is a whole number, else false
      */
     public static boolean isWhole(double n){
-        return (int)Math.floor(n)==(int)n;
+        return Math.floor(n) ==n;
     }
 
     /**
